Revert "fix test"
This reverts commit 0b6f32c2a7fbae6c5bc518bd06baf9581a0e9f91.

+++ src/aqi_updater/aq_sampling.py
@@ -52,7 +52,7 @@ def sample_aq_to_point_gdf(
         The name of the exported csv file (e.g. aqi_2019-11-08T14.csv).
     """
     gdf = sampling_gdf.copy()
-    aqi_raster = rasterio.open(aq_tif_file, dtype='float64')
+    aqi_raster = rasterio.open(aq_tif_file)
     # get coordinates of edge centers as list of tuples
     coords = [
         (x, y) for x, y
